+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [v5.11.0](https://github.com/auth0/react-native-auth0/tree/v5.11.0) (2026-07-31)
+
+[Full Changelog](https://github.com/auth0/react-native-auth0/compare/v5.10.0...v5.11.0)
+
+**Added**
+
+- feat: enforce IPSIE session_expiry with a SESSION_EXPIRED error [\#1597](https://github.com/auth0/react-native-auth0/pull/1597) ([subhankarmaiti](https://github.com/subhankarmaiti))
+- feat: add passkeys support for web [\#1604](https://github.com/auth0/react-native-auth0/pull/1604) ([NandanPrabhu](https://github.com/NandanPrabhu))
+- feat: add My Account API support on the web platform [\#1608](https://github.com/auth0/react-native-auth0/pull/1608) ([subhankarmaiti](https://github.com/subhankarmaiti))
+- feat: add actor token support for Custom Token Exchange impersonation & delegation [\#1595](https://github.com/auth0/react-native-auth0/pull/1595) ([subhankarmaiti](https://github.com/subhankarmaiti))
+- feat: add INVALID_CALLBACK_URL web auth error code [\#1603](https://github.com/auth0/react-native-auth0/pull/1603) ([subhankarmaiti](https://github.com/subhankarmaiti))
+
 ## [v5.10.0](https://github.com/auth0/react-native-auth0/tree/v5.10.0) (2026-07-17)
 
 [Full Changelog](https://github.com/auth0/react-native-auth0/compare/v5.9.0...v5.10.0)